Change  the element that will have divs added into
Earlier I had added the sixteen divisions to the body element but the
instructions just stated to add them to a div that should be created in
our html instead. This way we can style the division that holds the
other divisions and not style the body of the entire page.