Add styling to each div generated, expand parents
By specifically targeting the generated divisions from the javascript
by using attribute selectors, we can style just the divisions that we
generated.
By setting the body height and width to 100vw and vh we state that we
want to use the entire viewport for the body.
Likewise we state in the container id division to use 100 percent of
the width and height that are offered to it by its parent the body
element.
And finally the flex items of the container id value division will be
equally set to 6% for now, as 16 goes into 100 6.25 times.