Add reset, change values of generated divs
I added a reset to overwrite default styling as it was confusing my
math when calculating the area of each generated division.
I added flexbox to the body so that I could position the container
after I shrunk it so that it didn't take up all of the pages space.
As mentioned I changed the sizes of the generated divisions and also
the size of the container and as of right now the container uses a
default value of 480px for width and height as that would create a
sixteen by sixteen grid of 30px divisions.