Add button styling, change previous styling
The body actually needed to be changed to column flex direction and
have a align-items of center to horizontally center the container and
button appropriately.
The button's width was defined as 5% as it was taking 100% available to
it, but I'll probably change this later as with smaller viewports the
text  overflows from it.