Change createDivisions and event listener funcs
I had to change createDivisions function to accept not a count of total
divisions but a count of column/row's instead, as the execercise told
us to limit the user to 100 rows or columns, this was not implied
earlier.
I had to change the event listener logic to limit the input from the
user, I also set a default value so if they only press enter then the
original count of 16 columns/rows will be used.