Change createDivisions function, var and values
Inside the function we changed the name of the iteration division being
created to make more sense.
We also changed the width and height to be calculated using the
containers width divided by the number of rows/columns requested, so
now the generated divisions will fit inside the container and not flow
out as previously written.